fix: 수정 구매 금액 최대값 예외처리
리스트나 배열의 최대길이를 사용할 수 있도록 입력 금액을 Long으로 계산 후 변환

+++ src/main/java/lotto/Input.java
@@ -6,10 +6,10 @@ import java.util.List;
 import camp.nextstep.edu.missionutils.Console;
 
 class Input {
-    static Integer inputMoney() {
+    static Long inputMoney() {
         String raw = Console.readLine();
         Validation.inputMoneyExceptionCheck(raw);
-        return Integer.parseInt(raw);
+        return Long.parseLong(raw);
     }
 
     static Lotto inputWinnerNumber() {
